// src/common/enums/accessory-key.enum.ts
export enum AccessoryKey {
  BOTON_ENCENDIDO = 'boton_encendido',
  KIT_CARRETERA = 'kit_carretera',
  AROS = 'aros',
  LAMINAS = 'laminas',
  MOQUETAS = 'moquetas',
  CUBREMALETAS = 'cubremaletas',
  SEGURO = 'seguro',
  TELEMETRIA = 'telemetria',
  SENSORES = 'sensores',
  ALARMA = 'alarma',
  NEBLINEROS = 'neblineros',
  KIT_SEGURIDAD = 'kit_seguridad',
  PROTECTOR_CERAMICO = 'protector_ceramico',
  OTROS = 'otros',
}

export enum AccessoryClassification {
  VENDIDO = 'VENDIDO',
  OBSEQUIADO = 'OBSEQUIADO',
  NO_APLICA = 'NO_APLICA',
}